restructured
restructured taskModal and screens for task.
