+++ household-app/component/taskModal/addTaskModal/styles.ts
@@ -55,11 +55,10 @@ const styles = StyleSheet.create({
         backgroundColor: "rgba(0,0,0,0.5)",
     },
     modalView: {
-        display: "flex",
         marginLeft: 20,
         marginRight: 20,
         marginBottom: 20,
-        // backgroundColor: "#f2f2f2",
+        backgroundColor: "#f2f2f2",
         height: "80%",
         width: "95%",
         borderRadius: 20,
@@ -73,17 +72,6 @@ const styles = StyleSheet.create({
         shadowOpacity: 0.25,
         shadowRadius: 3.84,
         elevation: 5,
-    },
-    openButton: {
-        backgroundColor: "#F194FF",
-        borderRadius: 20,
-        padding: 10,
-        elevation: 2,
-    },
-    textStyle: {
-        // color: "white",
-        fontWeight: "bold",
-        textAlign: "center",
     },
     modalTextView: {
         flex: 1,
@@ -104,58 +92,6 @@ const styles = StyleSheet.create({
         fontSize: 24,
         fontWeight: "bold",
     },
-    container: {
-        marginTop: 15,
-    },
-    textinputTitleRight2: {
-        flex: 1,
-        justifyContent: "center",
-        alignContent: "flex-start",
-        top: 55,
-        marginLeft: 15,
-        marginTop: 0,
-        marginRight: 15,
-        marginBottom: 5,
-        maxHeight: 160,
-        minWidth: 360,
-        borderRadius: 8,
-        backgroundColor: "#ffffff",
-        shadowColor: "#000",
-        shadowOffset: {
-            width: 0,
-            height: 0,
-        },
-        shadowOpacity: 0.5,
-        shadowRadius: 12.35,
-        elevation: 10,
-    },
-    titleContent2: {
-        // backgroundColor: "white",
-        height: 166,
-    },
-    textinputTitleRight: {
-        flex: 1,
-        justifyContent: "center",
-        alignContent: "flex-end",
-        top: 55,
-        marginLeft: 15,
-        marginTop: 0,
-        marginRight: 15,
-        marginBottom: 0,
-        maxHeight: 60,
-        minWidth: 360,
-        shadowColor: "#000",
-        shadowOffset: {
-            width: 0,
-            height: 0,
-        },
-        shadowOpacity: 0.5,
-        shadowRadius: 12.35,
-        elevation: 10,
-    },
-    titleContent: {
-        // backgroundColor: "white",
-    },
     buttonsContainer: {
         alignItems: "center",
         flexDirection: "row",
@@ -173,7 +109,7 @@ const styles = StyleSheet.create({
         width: "50%",
         alignItems: "center",
         flexDirection: "row",
-        justifyContent: "space-evenly",
+        justifyContent: "center",
         shadowColor: "rgba(0, 0, 0, 0.1)",
         shadowOpacity: 0.8,
         elevation: 6,
@@ -190,7 +126,7 @@ const styles = StyleSheet.create({
         width: "50%",
         alignItems: "center",
         flexDirection: "row",
-        justifyContent: "space-evenly",
+        justifyContent: "center",
         shadowColor: "rgba(0, 0, 0, 0.1)",
         shadowOpacity: 0.8,
         elevation: 6,
@@ -198,7 +134,7 @@ const styles = StyleSheet.create({
         shadowOffset: { width: 1, height: 13 },
         borderBottomLeftRadius: 20,
     },
-    householdButtonText: {
+    taskButtonText: {
         // color: "black",
         fontSize: 18,
         fontWeight: "bold",
@@ -260,9 +196,6 @@ const styles = StyleSheet.create({
         shadowRadius: 12.35,
         elevation: 10,
     },
-    input2TextColor: {
-        color: "gray",
-    },
     clickedDay: {
         alignItems: "center",
         justifyContent: "space-between",
@@ -292,7 +225,6 @@ const styles = StyleSheet.create({
         // color: "black",
         fontSize: 18,
         fontWeight: "bold",
-        // padding: 12,
     },
     circleButton: {
         width: 36,
@@ -326,139 +258,6 @@ const styles = StyleSheet.create({
         fontSize: 14,
         justifyContent: "center",
         textAlign: "center",
-    },
-    householdButton: {
-        // backgroundColor: "white",
-        paddingVertical: 10,
-        paddingHorizontal: 10,
-        width: "50%",
-        alignItems: "center",
-        flexDirection: "row",
-        justifyContent: "center",
-        shadowColor: "rgba(0, 0, 0, 0.1)",
-        shadowOpacity: 0.8,
-        elevation: 6,
-        shadowRadius: 15,
-        shadowOffset: { width: 1, height: 13 },
-        borderBottomWidth: 1,
-        borderBottomColor: "gainsboro",
-    },
-    householdButton2: {
-        // backgroundColor: "white",
-        paddingVertical: 10,
-        paddingHorizontal: 10,
-        width: "50%",
-        alignItems: "center",
-        flexDirection: "row",
-        justifyContent: "center",
-        shadowColor: "rgba(0, 0, 0, 0.1)",
-        shadowOpacity: 0.8,
-        elevation: 6,
-        shadowRadius: 15,
-        shadowOffset: { width: 1, height: 13 },
-        borderStartWidth: 1,
-        borderBottomWidth: 1,
-        borderBottomColor: "gainsboro",
-        borderStartColor: "gainsboro",
-    },
-    modalView2: {
-        // margin: 20,
-        width: 300,
-        height: 500,
-        // backgroundColor: "#f2f2f2",
-        borderRadius: 20,
-        padding: 20,
-        alignItems: "center",
-        shadowColor: "#000",
-        shadowOffset: {
-            width: 0,
-            height: 2,
-        },
-        shadowOpacity: 0.25,
-        shadowRadius: 3.84,
-        elevation: 5,
-    },
-    modalViewDelete: {
-        // margin: 20,
-        width: 400,
-        height: 200,
-        // backgroundColor: "#f2f2f2",
-        borderRadius: 20,
-        padding: 20,
-        alignItems: "center",
-        shadowColor: "#000",
-        shadowOffset: {
-            width: 0,
-            height: 2,
-        },
-        shadowOpacity: 0.25,
-        shadowRadius: 3.84,
-        elevation: 5,
-    },
-    taskText: {
-        fontSize: 16,
-        marginBottom: 10,
-    },
-    modalText2: {
-        // marginBottom: 15,
-        // textAlign: "center",
-        fontWeight: "bold",
-        fontSize: 18,
-    },
-    saveButtonDelete: {
-        // backgroundColor: "white",
-        paddingVertical: 20,
-        paddingHorizontal: 20,
-        width: "33.3%",
-        alignItems: "center",
-        flexDirection: "row",
-        justifyContent: "center",
-        shadowColor: "rgba(0, 0, 0, 0.1)",
-        shadowOpacity: 0.8,
-        elevation: 6,
-        shadowRadius: 15,
-        shadowOffset: { width: 1, height: 13 },
-        borderBottomLeftRadius: 20,
-    },
-    archiveButton: {
-        // backgroundColor: "white",
-        paddingVertical: 20,
-        paddingHorizontal: 20,
-        width: "33.3%",
-        alignItems: "center",
-        flexDirection: "row",
-        justifyContent: "center",
-        shadowColor: "rgba(0, 0, 0, 0.1)",
-        shadowOpacity: 0.8,
-        elevation: 6,
-        shadowRadius: 15,
-        shadowOffset: { width: 1, height: 13 },
-        borderStartWidth: 1,
-        borderStartColor: "gainsboro",
-    },
-    closeButtonDelete: {
-        // backgroundColor: "white",
-        paddingVertical: 20,
-        paddingHorizontal: 20,
-        width: "33.3%",
-        alignItems: "center",
-        flexDirection: "row",
-        justifyContent: "center",
-        shadowColor: "rgba(0, 0, 0, 0.1)",
-        shadowOpacity: 0.8,
-        elevation: 6,
-        shadowRadius: 15,
-        shadowOffset: { width: 1, height: 13 },
-        borderBottomRightRadius: 20,
-        borderStartWidth: 1,
-        borderStartColor: "gainsboro",
-    },
-    warningText: {
-        marginBottom: 15,
-        textAlign: "center",
-        fontWeight: "bold",
-        fontSize: 20,
-        color: "red",
     },
 });
