+++ household-app/component/taskModal/editTaskInputModal/styles.ts
@@ -55,10 +55,11 @@ const styles = StyleSheet.create({
         backgroundColor: "rgba(0,0,0,0.5)",
     },
     modalView: {
+        display: "flex",
         marginLeft: 20,
         marginRight: 20,
         marginBottom: 20,
-        backgroundColor: "#f2f2f2",
+        // backgroundColor: "#f2f2f2",
         height: "80%",
         width: "95%",
         borderRadius: 20,
@@ -72,17 +73,6 @@ const styles = StyleSheet.create({
         shadowOpacity: 0.25,
         shadowRadius: 3.84,
         elevation: 5,
-    },
-    openButton: {
-        backgroundColor: "#F194FF",
-        borderRadius: 20,
-        padding: 10,
-        elevation: 2,
-    },
-    textStyle: {
-        color: "white",
-        fontWeight: "bold",
-        textAlign: "center",
     },
     modalTextView: {
         flex: 1,
@@ -103,58 +93,6 @@ const styles = StyleSheet.create({
         fontSize: 24,
         fontWeight: "bold",
     },
-    container: {
-        marginTop: 15,
-    },
-    textinputTitleRight2: {
-        flex: 1,
-        justifyContent: "center",
-        alignContent: "flex-start",
-        top: 55,
-        marginLeft: 15,
-        marginTop: 0,
-        marginRight: 15,
-        marginBottom: 5,
-        maxHeight: 160,
-        minWidth: 360,
-        borderRadius: 8,
-        backgroundColor: "#ffffff",
-        shadowColor: "#000",
-        shadowOffset: {
-            width: 0,
-            height: 0,
-        },
-        shadowOpacity: 0.5,
-        shadowRadius: 12.35,
-        elevation: 10,
-    },
-    titleContent2: {
-        backgroundColor: "white",
-        height: 166,
-    },
-    textinputTitleRight: {
-        flex: 1,
-        justifyContent: "center",
-        alignContent: "flex-end",
-        top: 55,
-        marginLeft: 15,
-        marginTop: 0,
-        marginRight: 15,
-        marginBottom: 0,
-        maxHeight: 60,
-        minWidth: 360,
-        shadowColor: "#000",
-        shadowOffset: {
-            width: 0,
-            height: 0,
-        },
-        shadowOpacity: 0.5,
-        shadowRadius: 12.35,
-        elevation: 10,
-    },
-    titleContent: {
-        backgroundColor: "white",
-    },
     buttonsContainer: {
         alignItems: "center",
         flexDirection: "row",
@@ -172,7 +110,7 @@ const styles = StyleSheet.create({
         width: "50%",
         alignItems: "center",
         flexDirection: "row",
-        justifyContent: "center",
+        justifyContent: "space-evenly",
         shadowColor: "rgba(0, 0, 0, 0.1)",
         shadowOpacity: 0.8,
         elevation: 6,
@@ -189,19 +127,13 @@ const styles = StyleSheet.create({
         width: "50%",
         alignItems: "center",
         flexDirection: "row",
-        justifyContent: "center",
+        justifyContent: "space-evenly",
         shadowColor: "rgba(0, 0, 0, 0.1)",
         shadowOpacity: 0.8,
         elevation: 6,
         shadowRadius: 15,
         shadowOffset: { width: 1, height: 13 },
         borderBottomLeftRadius: 20,
-    },
-    householdButtonText: {
-        // color: "black",
-        fontSize: 18,
-        fontWeight: "bold",
-        marginLeft: 15,
     },
     inputsCard: {
         // flex: 1,
@@ -259,9 +191,6 @@ const styles = StyleSheet.create({
         shadowRadius: 12.35,
         elevation: 10,
     },
-    input2TextColor: {
-        color: "gray",
-    },
     clickedDay: {
         alignItems: "center",
         justifyContent: "space-between",
@@ -291,6 +220,7 @@ const styles = StyleSheet.create({
         // color: "black",
         fontSize: 18,
         fontWeight: "bold",
+        // padding: 12,
     },
     circleButton: {
         width: 36,
