+++ household-app/component/taskModal/styles.ts
@@ -1,0 +1,245 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+    centeredView: {
+        flex: 1,
+        justifyContent: "center",
+        alignItems: "center",
+        marginTop: 22,
+    },
+    centeredViewBlurred: {
+        flex: 1,
+        justifyContent: "center",
+        alignItems: "center",
+        marginTop: 22,
+        backgroundColor: "rgba(0,0,0,0.5)",
+    },
+    modalView: {
+        display: "flex",
+        marginLeft: 20,
+        marginRight: 20,
+        marginBottom: 20,
+        // backgroundColor: "#f2f2f2",
+        height: "80%",
+        width: "95%",
+        borderRadius: 20,
+        padding: 35,
+        alignItems: "center",
+        shadowColor: "#000",
+        shadowOffset: {
+            width: 0,
+            height: 2,
+        },
+        shadowOpacity: 0.25,
+        shadowRadius: 3.84,
+        elevation: 5,
+    },
+    modalTextView: {
+        flex: 1,
+        alignItems: "flex-start",
+        position: "absolute",
+        top: 0,
+        left: 0,
+        right: 0,
+        minWidth: "100%",
+        paddingVertical: 20,
+        paddingHorizontal: 20,
+        borderTopLeftRadius: 20,
+        borderTopRightRadius: 20,
+        // backgroundColor: "white",
+    },
+    modalText: {
+        textAlign: "center",
+        fontSize: 24,
+        fontWeight: "bold",
+    },
+    buttonsContainer: {
+        alignItems: "center",
+        flexDirection: "row",
+        justifyContent: "center",
+        alignSelf: "flex-end",
+        position: "absolute",
+        bottom: 0,
+        left: 0,
+        right: 0,
+    },
+    closeButton: {
+        // backgroundColor: "white",
+        paddingVertical: 20,
+        paddingHorizontal: 20,
+        width: "50%",
+        alignItems: "center",
+        flexDirection: "row",
+        justifyContent: "space-evenly",
+        shadowColor: "rgba(0, 0, 0, 0.1)",
+        shadowOpacity: 0.8,
+        elevation: 6,
+        shadowRadius: 15,
+        shadowOffset: { width: 1, height: 13 },
+        borderBottomRightRadius: 20,
+        borderStartWidth: 1,
+        borderStartColor: "gainsboro",
+    },
+    saveButton: {
+        // backgroundColor: "white",
+        paddingVertical: 20,
+        paddingHorizontal: 20,
+        width: "50%",
+        alignItems: "center",
+        flexDirection: "row",
+        justifyContent: "space-evenly",
+        shadowColor: "rgba(0, 0, 0, 0.1)",
+        shadowOpacity: 0.8,
+        elevation: 6,
+        shadowRadius: 15,
+        shadowOffset: { width: 1, height: 13 },
+        borderBottomLeftRadius: 20,
+    },
+    taskButtonText: {
+        // color: "black",
+        fontSize: 18,
+        fontWeight: "bold",
+        marginLeft: 15,
+    },
+    buttonText: {
+        // color: "black",
+        fontSize: 18,
+        fontWeight: "bold",
+        // padding: 12,
+    },
+    taskButton: {
+        // backgroundColor: "white",
+        paddingVertical: 10,
+        paddingHorizontal: 10,
+        width: "50%",
+        alignItems: "center",
+        flexDirection: "row",
+        justifyContent: "center",
+        shadowColor: "rgba(0, 0, 0, 0.1)",
+        shadowOpacity: 0.8,
+        elevation: 6,
+        shadowRadius: 15,
+        shadowOffset: { width: 1, height: 13 },
+        borderBottomWidth: 1,
+        borderBottomColor: "gainsboro",
+    },
+    taskButton2: {
+        // backgroundColor: "white",
+        paddingVertical: 10,
+        paddingHorizontal: 10,
+        width: "50%",
+        alignItems: "center",
+        flexDirection: "row",
+        justifyContent: "center",
+        shadowColor: "rgba(0, 0, 0, 0.1)",
+        shadowOpacity: 0.8,
+        elevation: 6,
+        shadowRadius: 15,
+        shadowOffset: { width: 1, height: 13 },
+        borderStartWidth: 1,
+        borderBottomWidth: 1,
+        borderBottomColor: "gainsboro",
+        borderStartColor: "gainsboro",
+    },
+    modalView2: {
+        // margin: 20,
+        width: 300,
+        height: 500,
+        // backgroundColor: "#f2f2f2",
+        borderRadius: 20,
+        padding: 20,
+        alignItems: "center",
+        shadowColor: "#000",
+        shadowOffset: {
+            width: 0,
+            height: 2,
+        },
+        shadowOpacity: 0.25,
+        shadowRadius: 3.84,
+        elevation: 5,
+    },
+    modalViewDelete: {
+        // margin: 20,
+        width: 400,
+        height: 200,
+        // backgroundColor: "#f2f2f2",
+        borderRadius: 20,
+        padding: 20,
+        alignItems: "center",
+        shadowColor: "#000",
+        shadowOffset: {
+            width: 0,
+            height: 2,
+        },
+        shadowOpacity: 0.25,
+        shadowRadius: 3.84,
+        elevation: 5,
+    },
+    taskText: {
+        fontSize: 16,
+        marginBottom: 10,
+    },
+    modalText2: {
+        // marginBottom: 15,
+        // textAlign: "center",
+        fontWeight: "bold",
+        fontSize: 18,
+    },
+    saveButtonDelete: {
+        // backgroundColor: "white",
+        paddingVertical: 20,
+        paddingHorizontal: 20,
+        width: "33.3%",
+        alignItems: "center",
+        flexDirection: "row",
+        justifyContent: "center",
+        shadowColor: "rgba(0, 0, 0, 0.1)",
+        shadowOpacity: 0.8,
+        elevation: 6,
+        shadowRadius: 15,
+        shadowOffset: { width: 1, height: 13 },
+        borderBottomLeftRadius: 20,
+    },
+    archiveButton: {
+        // backgroundColor: "white",
+        paddingVertical: 20,
+        paddingHorizontal: 20,
+        width: "33.3%",
+        alignItems: "center",
+        flexDirection: "row",
+        justifyContent: "center",
+        shadowColor: "rgba(0, 0, 0, 0.1)",
+        shadowOpacity: 0.8,
+        elevation: 6,
+        shadowRadius: 15,
+        shadowOffset: { width: 1, height: 13 },
+        borderStartWidth: 1,
+        borderStartColor: "gainsboro",
+    },
+    closeButtonDelete: {
+        // backgroundColor: "white",
+        paddingVertical: 20,
+        paddingHorizontal: 20,
+        width: "33.3%",
+        alignItems: "center",
+        flexDirection: "row",
+        justifyContent: "center",
+        shadowColor: "rgba(0, 0, 0, 0.1)",
+        shadowOpacity: 0.8,
+        elevation: 6,
+        shadowRadius: 15,
+        shadowOffset: { width: 1, height: 13 },
+        borderBottomRightRadius: 20,
+        borderStartWidth: 1,
+        borderStartColor: "gainsboro",
+    },
+    warningText: {
+        marginBottom: 15,
+        textAlign: "center",
+        fontWeight: "bold",
+        fontSize: 20,
+        color: "red",
+    },
+});
+
+export default styles;
