+++ household-app/navigation/AuthStack.tsx
@@ -2,15 +2,15 @@ import React from "react";
 import { createNativeStackNavigator } from "@react-navigation/native-stack";
 import LoginScreen from "../screens/Account/LoginScreen";
 import ProfileScreen from "../screens/Profile/ProfileScreen";
-import TasksScreen from "../screens/Tasks/TasksScreen";
+import TasksScreen from "../screens/Tasks/tasksScreen/TasksScreen";
 import { createMaterialTopTabNavigator } from "@react-navigation/material-top-tabs";
 import { MainRoutes, MainStack } from "../routes/routes";
 import CreateAccountScreen from "../screens/Account/CreateAccount";
 import { MaterialIcons } from "@expo/vector-icons";
 import { Button, Pressable } from "react-native";
 import { useSafeAreaInsets } from "react-native-safe-area-context";
-import LastWeekScreen from "../screens/Tasks/LastWeekScreen";
-import LastMonthScreen from "../screens/Tasks/LastMonthScreen";
+import LastWeekScreen from "../screens/Tasks/lastWeekScreen/LastWeekScreen";
+import LastMonthScreen from "../screens/Tasks/lastMonthScreen/LastMonthScreen";
 import HouseholdScreen from "../screens/Household/HouseholdScreen";
 import Constants from "expo-constants";
