+++ household-app/screens/Tasks/lastMonthScreen/styles.ts
@@ -1,0 +1,14 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+    container: {
+        flex: 1,
+        alignItems: "center",
+        justifyContent: "center",
+    },
+    text: {
+        flex: 1,
+    },
+});
+
+export default styles;
