+++ household-app/screens/Tasks/lastWeekScreen/LastWeekScreen.tsx
@@ -2,19 +2,20 @@
 /* eslint-disable @typescript-eslint/no-non-null-assertion */
 
 import React, { FC } from "react";
-import { SafeAreaView, ScrollView, StyleSheet, View } from "react-native";
+import { SafeAreaView, ScrollView, View } from "react-native";
 import { Text } from "react-native-paper";
-import StatisticsCharts from "../../component/piecharts/StatisticsCharts";
-import { selectSelectedHousehold } from "../../Redux/features/SelectedState/SelectedStateSelectors";
-import { useAppSelector } from "../../Redux/hooks";
-import { useGetDoneTasksWithHouseholdIdQuery } from "../../Redux/Service/doneTask/doneTaskApi";
-import { FeedStackScreenProps, MainRoutes } from "../../routes/routes";
-import { getCalendarWeekDoneTasksByHousehold } from "./doneTaskHelper";
-import { createMemberStatistics, MemberStatistics } from "./MemberStatistics";
+import StatisticsCharts from "../../../component/piecharts/StatisticsCharts";
+import { selectSelectedHousehold } from "../../../Redux/features/SelectedState/SelectedStateSelectors";
+import { useAppSelector } from "../../../Redux/hooks";
+import { useGetDoneTasksWithHouseholdIdQuery } from "../../../Redux/Service/doneTask/doneTaskApi";
+import { FeedStackScreenProps, MainRoutes } from "../../../routes/routes";
+import { getCalendarWeekDoneTasksByHousehold } from "../helpers/doneTaskHelper";
+import { createMemberStatistics, MemberStatistics } from "../helpers/MemberStatistics";
+import styles from "./styles";
 
 type Props = FeedStackScreenProps<MainRoutes.ProfileScreen>;
 
-const CurrentWeekScreen: FC<Props> = ({ navigation }: Props): React.ReactElement => {
+const LastWeekScreen: FC<Props> = ({ navigation }: Props): React.ReactElement => {
     const currentHousehold = useAppSelector(selectSelectedHousehold);
     const { data: doneTasksArray, error: doneTaskError } = useGetDoneTasksWithHouseholdIdQuery(currentHousehold?.id!);
 
@@ -22,7 +23,7 @@ const CurrentWeekScreen: FC<Props> = ({ navigation }: Props): React.ReactElement
     let fillerMessage = "No done tasks found for this household.";
 
     if (doneTasksArray !== undefined && doneTasksArray.length > 0 && currentHousehold !== undefined) {
-        const doneTasksOfLastWeek = getCalendarWeekDoneTasksByHousehold(doneTasksArray, currentHousehold, 0);
+        const doneTasksOfLastWeek = getCalendarWeekDoneTasksByHousehold(doneTasksArray, currentHousehold, 1);
         if (doneTasksOfLastWeek.length === 0) {
             fillerMessage = "No data found for the selected period.";
         } else {
@@ -45,15 +46,4 @@ const CurrentWeekScreen: FC<Props> = ({ navigation }: Props): React.ReactElement
     );
 };
 
-export default CurrentWeekScreen;
-
-const styles = StyleSheet.create({
-    container: {
-        flex: 1,
-        alignItems: "center",
-        justifyContent: "center",
-    },
-    text: {
-        flex: 1,
-    },
-});
+export default LastWeekScreen;
