+++ household-app/screens/Tasks/usersInHousehold/UsersInHouseHoldScreen.tsx
@@ -1,20 +1,21 @@
 /* eslint-disable react/jsx-no-undef */
 import React, { FC, useContext, useEffect, useState } from "react";
-import { Dimensions, FlatList, StyleSheet, View } from "react-native";
+import { FlatList, View } from "react-native";
 import { Surface, Text } from "react-native-paper";
-import Button from "../../component/common/Button";
-import ChangeHouseholdNameModal from "../../component/householdComponents/changeHouseholdNameModal/changeHouseholdNameModal";
-import ChangeMemberStatusModal from "../../component/householdComponents/changeMemberStatusModal/changeMemberStatusModal";
-import LeaveModal from "../../component/householdComponents/leaveModal/leaveModal";
-import PendingMemberTaskCard from "../../component/householdComponents/pendingMemberCard/pendingMemberCard";
-import SnackbarComponent from "../../component/snackbar/snackbarComponent";
-import UserListComponent from "../../component/taskFolder/householdComponent";
-import { snackbarContext } from "../../context/snackBarContext";
-import { selectCurrentLoginUser } from "../../Redux/features/loginUser/LoginSelectors";
-import { selectSelectedHousehold } from "../../Redux/features/SelectedState/SelectedStateSelectors";
-import { useAppSelector } from "../../Redux/hooks";
-import { useLazyGetHouseholdByIdQuery, useLeaveHouseholdMutation } from "../../Redux/Service/household/householdApi";
-import { FeedStackScreenProps, MainRoutes } from "../../routes/routes";
+import Button from "../../../component/common/Button";
+import ChangeHouseholdNameModal from "../../../component/householdComponents/changeHouseholdNameModal/changeHouseholdNameModal";
+import ChangeMemberStatusModal from "../../../component/householdComponents/changeMemberStatusModal/changeMemberStatusModal";
+import LeaveModal from "../../../component/householdComponents/leaveModal/leaveModal";
+import PendingMemberTaskCard from "../../../component/householdComponents/pendingMemberCard/pendingMemberCard";
+import SnackbarComponent from "../../../component/snackbar/snackbarComponent";
+import UserListComponent from "../../../component/taskFolder/householdComponent";
+import { snackbarContext } from "../../../context/snackBarContext";
+import { selectCurrentLoginUser } from "../../../Redux/features/loginUser/LoginSelectors";
+import { selectSelectedHousehold } from "../../../Redux/features/SelectedState/SelectedStateSelectors";
+import { useAppSelector } from "../../../Redux/hooks";
+import { useLazyGetHouseholdByIdQuery, useLeaveHouseholdMutation } from "../../../Redux/Service/household/householdApi";
+import { FeedStackScreenProps, MainRoutes } from "../../../routes/routes";
+import styles from "./styles";
 
 type Props = FeedStackScreenProps<MainRoutes.UsersInHouseHoldScreen>;
 const UsersInHouseHoldScreen: FC<Props> = ({ navigation }: Props): React.ReactElement => {
@@ -175,104 +176,6 @@ const UsersInHouseHoldScreen: FC<Props> = ({ navigation }: Props): React.ReactEl
 };
 
 export default UsersInHouseHoldScreen;
-
-const deviceHeight = Math.round(Dimensions.get("window").height);
-
-const styles = StyleSheet.create({
-    listContainer: {
-        maxHeight: deviceHeight - 241,
-    },
-    container: {
-        flex: 1,
-    },
-    text: {
-        color: "grey",
-    },
-    card: {
-        flexDirection: "row",
-        shadowOffset: { width: 5, height: 5 },
-        width: "90%",
-        borderRadius: 12,
-        alignSelf: "center",
-        marginTop: 5,
-        marginBottom: 5,
-    },
-    householdButton: {
-        backgroundColor: "white",
-        paddingVertical: 15,
-        paddingHorizontal: 15,
-        width: "45%",
-        alignItems: "center",
-        flexDirection: "row",
-        justifyContent: "center",
-        shadowColor: "rgba(0, 0, 0, 0.1)",
-        shadowOpacity: 0.8,
-        elevation: 6,
-        shadowRadius: 15,
-        shadowOffset: { width: 1, height: 13 },
-        borderRadius: 20,
-        marginBottom: 15,
-        marginLeft: 10,
-        marginRight: 10,
-        height: 55,
-    },
-    inviteCode: {
-        textAlign: "center",
-        alignSelf: "center",
-        width: "100%",
-        height: 30,
-        // color: "red",
-    },
-    householdButtonUser: {
-        backgroundColor: "white",
-        paddingVertical: 15,
-        paddingHorizontal: 15,
-        width: "45%",
-        alignItems: "center",
-        flexDirection: "row",
-        justifyContent: "center",
-        shadowColor: "rgba(0, 0, 0, 0.1)",
-        shadowOpacity: 0.8,
-        elevation: 6,
-        shadowRadius: 15,
-        shadowOffset: { width: 1, height: 13 },
-        borderRadius: 20,
-        marginBottom: 15,
-        marginLeft: 10,
-        marginRight: 10,
-        height: 55,
-    },
-    buttonText: {
-        color: "grey",
-        fontSize: 16,
-    },
-    householdButtonText: {
-        color: "black",
-        fontSize: 18,
-        fontWeight: "bold",
-        marginLeft: 10,
-    },
-    buttonsContainer: {
-        alignItems: "center",
-        flexDirection: "row",
-        justifyContent: "space-between",
-        alignSelf: "flex-end",
-        position: "absolute",
-        bottom: 0,
-        left: 0,
-        right: 0,
-    },
-    buttonsContainerUser: {
-        alignItems: "center",
-        flexDirection: "row",
-        justifyContent: "center",
-        alignSelf: "flex-end",
-        position: "absolute",
-        bottom: 0,
-        left: 0,
-        right: 0,
-    },
-});
 
 interface fullMemberInfo {
     name: string;
